CP working

+++ plane.py
@@ -5,10 +5,10 @@ from dataclasses import dataclass
 class Plane:
     id: int
     airland_id: int
-    A: float    # appearence time
-    E: float    # earliest time
-    T: float    # target time
-    L: float    # latest time
-    PCb: float  # cost per unit of time for arriving before target time
-    PCa: float  # cost per unit of time for arriving after target time
-    actual_land_time: float = 0.0
+    A: int    # appearence time
+    E: int    # earliest time
+    T: int    # target time
+    L: int    # latest time
+    PCb: int  # cost per unit of time for arriving before target time
+    PCa: int  # cost per unit of time for arriving after target time
+    actual_land_time: int = 0
